continued with jdbcinputexample
